requirements.txt
Flask
pyjwt
requests
oauthlib
cachetools
setuptools
requests_oauthlib
